Move InvalidParameterException to correct package
