+++ server/src/main/java/com/ibm/ws/lars/rest/AssetQueryParameters.java
@@ -32,6 +32,7 @@ import javax.ws.rs.core.MultivaluedMap;
 import javax.ws.rs.core.UriInfo;
 
 import com.ibm.ws.lars.rest.Condition.Operation;
+import com.ibm.ws.lars.rest.exceptions.InvalidParameterException;
 import com.ibm.ws.lars.rest.exceptions.RepositoryException;
 
 /**
