+++ server/src/main/java/com/ibm/ws/lars/rest/RepositoryRESTResource.java
@@ -58,6 +58,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.ibm.ws.lars.rest.exceptions.AssetPersistenceException;
 import com.ibm.ws.lars.rest.exceptions.InvalidIdException;
 import com.ibm.ws.lars.rest.exceptions.InvalidJsonAssetException;
+import com.ibm.ws.lars.rest.exceptions.InvalidParameterException;
 import com.ibm.ws.lars.rest.exceptions.NonExistentArtefactException;
 import com.ibm.ws.lars.rest.exceptions.RepositoryException;
 import com.ibm.ws.lars.rest.model.Asset;
